requirements.txt
Flask==3.0.3
SQLAlchemy==2.0.36
psycopg[binary]
requests==2.32.3
python-dotenv==1.0.1
APScheduler==3.10.4
pytz==2024.1
gunicorn==21.2.0
